# docs/reference/solar_complete_spec_md/07_openapi.yaml
openapi: 3.1.0
info:
  title: Kuozui Green Energy Display API
  version: 1.0.0
  description: API for solar green energy display player using Fastify, MQTT, Socket.IO, and SQLite.
servers:
  - url: /api
paths:
...
  /images:
    get:
      summary: Get image assets
      operationId: getImages
      responses:
        '200':
          description: Image assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  images:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImageAsset'
    post:
      summary: Upload image
      operationId: uploadImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Uploaded image
  /images/{id}:
    put:
      summary: Update image metadata
      operationId: updateImage
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageAssetInput'
      responses:
        '200':
          description: Updated image
    delete:
      summary: Delete image
      operationId: deleteImage
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Deleted image
  /images/reorder:
    put:
      summary: Reorder images
      operationId: reorderImages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Reordered
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ImageAsset:
      allOf:
        - $ref: '#/components/schemas/ImageAssetInput'
        - type: object
          properties:
            id: { type: string }
            filename: { type: string }
            url: { type: string }
            width: { type: integer }
            height: { type: integer }
            fileSize: { type: integer }
            createdAt: { type: string, format: date-time }
    ImageAssetInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        displayOrder: { type: integer }
        durationSec: { type: integer }
        includeInSlideshow: { type: boolean }
        isCover: { type: boolean }
        enabled: { type: boolean }
...